+++ R/data.R
@@ -5,7 +5,7 @@
 #' hex code.
 #'
 #' @format A character vector
-#' @source \href{https://analysisfunction.civilservice.gov.uk/policy-store/data-visualisation-colours-in-charts/}{Government Analysis Function Colours Guidance}
+#' @source \href{https://designsystem.gov.scot/guidance/charts/data-visualisation-colour-palettes}{Scottish Government Design System}
 # nolint end
 
 "sg_colour_values"
@@ -17,7 +17,32 @@
 #' @description A list grouping colours into palettes.
 #'
 #' @format A character list
-#' @source \href{https://analysisfunction.civilservice.gov.uk/policy-store/data-visualisation-colours-in-charts/}{Government Analysis Function Colours Guidance}
+#' @source \href{https://designsystem.gov.scot/guidance/charts/data-visualisation-colour-palettes}{Scottish Government Design System}
 # nolint end
 
 "sg_colour_palettes"
+
+
+# nolint start
+#' @title Analysis Function colour names and hex codes
+#'
+#' @description A vector containing colour names and their corresponding
+#' hex code.
+#'
+#' @format A character vector
+#' @source \href{https://analysisfunction.civilservice.gov.uk/policy-store/data-visualisation-colours-in-charts/}{Government Analysis Function Colours Guidance}
+# nolint end
+
+"af_colour_values"
+
+
+# nolint start
+#' @title Analysis Function colour palettes
+#'
+#' @description A list grouping colours into palettes.
+#'
+#' @format A character list
+#' @source \href{https://analysisfunction.civilservice.gov.uk/policy-store/data-visualisation-colours-in-charts/}{Government Analysis Function Colours Guidance}
+# nolint end
+
+"af_colour_palettes"