+++ R/scale_colour_continuous_sg.R
@@ -1,7 +1,9 @@
 #' @title Continuous colour scales for Scottish Government plots
 #'
-#' @param palette Name of palette to use from `sg_colour_palettes`; e.g. main,
-#' sequential, focus. Default value is 'sequential'.
+#' @param palette Name of palette to use; e.g. "main", "sequential", "focus".
+#' Default value is "sequential".
+#' @param palette_type Either "sg" to use Scottish Government palettes, or "af"
+#' to use Analysis Function palettes. Defaults to "sg".
 #' @param reverse Boolean value to indicate whether the palette should be
 #' reversed.
 #' @param na_colour Colour to set for missing values.
@@ -18,22 +20,30 @@
 #' @export
 
 scale_colour_continuous_sg <- function(palette = "sequential",
+                                       palette_type = c("sg", "af"),
                                        reverse = FALSE,
                                        na_colour = "grey50",
                                        guide = "colourbar",
                                        ...) {
+
+  palette_type <- match.arg(palette_type)
+
+  palette_list <- get(paste0(palette_type, "_colour_palettes"),
+                      as.environment("package:sgplot"))
+
   # Error if palette doesn't exist
-  if (!palette %in% names(sgplot::sg_colour_palettes)) {
+  if (!palette %in% names(palette_list)) {
     cli::cli_abort(c(
-      "x" = paste(palette, "is not a valid palette name.")
+      "x" = paste("`{palette}` is not a valid palette name in",
+                  "`{palette_type}_colour_palettes`.")
     ))
   }
 
-  colours <- as.vector(sgplot::sg_colour_palettes[[palette]])
+  colours <- as.vector(palette_list[[palette]])
 
   ggplot2::continuous_scale(
     aesthetics = "colour",
-    scale_name = "sg_continuous",
+    scale_name = paste0(palette_type, "_continuous"),
     palette    = scales::gradient_n_pal(colours, values = NULL, "Lab"),
     na.value   = na_colour,
     guide      = guide,