+++ R/scale_fill_discrete_sg.R
@@ -1,7 +1,9 @@
 #' @title Discrete colour fill scales for Scottish Government plots
 #'
-#' @param palette Name of palette to use from `sg_colour_palettes`; e.g. main,
-#' sequential, focus. Default value is 'main'.
+#' @param palette Name of palette to use; e.g. "main", "sequential", "focus."
+#' Default value is "main."
+#' @param palette_type Either "sg" to use Scottish Government palettes, or "af"
+#' to use Analysis Function palettes. Defaults to "sg".
 #' @param reverse Boolean value to indicate whether the palette should be
 #' reversed.
 #' @param ... Additional arguments passed to scale type.
@@ -18,11 +20,15 @@
 #' @export
 
 scale_fill_discrete_sg <- function(palette = "main",
+                                   palette_type = c("sg", "af"),
                                    reverse = FALSE,
                                    ...) {
+
+  palette_type <- match.arg(palette_type)
+
   ggplot2::discrete_scale("fill",
-    paste0("sg_", palette),
-    palette = sg_palette(palette, reverse),
+    paste0(palette_type, "_", palette),
+    palette = sg_palette(palette, reverse, palette_type = palette_type),
     ...
   )
 }