+++ data-raw/sg_colours.R
@@ -1,28 +1,27 @@
-# This script defines the colours and colour palettes for use in Scottish
-# Government plots.
-#
-# Colours are taken from Government Analysis Function guidance:
-# https://analysisfunction.civilservice.gov.uk/policy-store/
-#   data-visualisation-colours-in-charts/
+# Scottish Government (SG) Design System colours and palettes ----
+# Source: https://designsystem.gov.scot/guidance/
+#   charts/data-visualisation-colour-palettes
 
 sg_colour_values <- c(
-  `dark-blue`    = "#12436D",
-  `turquoise`    = "#28A197",
-  `dark-pink`    = "#801650",
-  `orange`       = "#F46A25",
-  `dark-grey`    = "#3D3D3D",
-  `light-purple` = "#A285D1",
-  `mid-blue`     = "#2073BC",
-  `light-blue`   = "#6BACE6",
-  `grey`         = "#BFBFBF"
+  `dark-blue` = "#002d54",
+  `teal` = "#2b9c93",
+  `purple` = "#6a2063",
+  `orange` = "#e5682a",
+  `dark-green` = "#0b4c0b",
+  `green` = "#5d9f3c",
+  `brown` = "#592c20",
+  `pink` = "#ca72a2",
+  `mid-blue` = "#0065bd",
+  `light-blue` = "#55a8f2",
+  `grey` = "#949494"
 )
 
 sg_colour_palettes <- list(
-  `main` = sg_colour_values[c("dark-blue", "turquoise", "dark-pink", "orange")],
-  `main2` = sg_colour_values[c("dark-blue", "orange")],
-  `main6` = sg_colour_values[c(
-    "dark-blue", "turquoise", "dark-pink",
-    "orange", "dark-grey", "light-purple"
+  `main` = sg_colour_values[c("dark-blue", "teal", "purple", "orange")],
+  `main-extended` = sg_colour_values[c(
+    "dark-blue", "teal", "purple",
+    "orange", "dark-green", "green",
+    "brown", "pink"
   )],
   `sequential` = sg_colour_values[c("dark-blue", "mid-blue", "light-blue")],
   `focus` = sg_colour_values[c("dark-blue", "grey")]